Email/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
APScheduler==3.10.4
